requirements.txt
opencv-contrib-python
Pillow
numpy
streamlit-webrtc
streamlit